+++ Source/Sitecore.FakeDb.Tests/GettingStarted.cs
@@ -63,19 +63,16 @@
     [Fact]
     public void HowDoICreateAnItemOfSpecificTemplate()
     {
-      // arrange
-      Sitecore.Data.TemplateID templateId = new Sitecore.Data.TemplateID(Sitecore.Data.ID.NewID);
+      Sitecore.Data.ID templateId = Sitecore.Data.ID.NewID;
 
-      // act
       using (Sitecore.FakeDb.Db db = new Sitecore.FakeDb.Db
         {
           new Sitecore.FakeDb.DbTemplate("products", templateId) { "Name" },
-          new Sitecore.FakeDb.DbItem("Apple", templateId)
+          new Sitecore.FakeDb.DbItem("Apple") { TemplateID = templateId }
         })
       {
-        // assert
         Sitecore.Data.Items.Item item = db.GetItem("/sitecore/content/apple");
-        item.TemplateID.Should().Be(templateId.ID);
+        item.TemplateID.Should().Be(templateId);
         item.Fields["Name"].Should().NotBeNull();
       }
     }
@@ -87,13 +84,11 @@
     [Fact]
     public void HowDoIConfigureItemAccess()
     {
-      // arrange & act
       using (Sitecore.FakeDb.Db db = new Sitecore.FakeDb.Db
         {
           new Sitecore.FakeDb.DbItem("home") { Access = { CanRead = false } }
         })
       {
-        // assert
         Sitecore.Data.Items.Item item = db.GetItem("/sitecore/content/home");
         item.Should().BeNull();
       }
@@ -121,7 +116,6 @@
     [Fact]
     public void HowDoIMockContentSearchLogic()
     {
-      // arrange
       try
       {
         var index = Substitute.For<Sitecore.ContentSearch.ISearchIndex>();
@@ -133,10 +127,8 @@
           searchResultItem.GetItem().Returns(db.GetItem("/sitecore/content/home"));
           index.CreateSearchContext().GetQueryable<Sitecore.ContentSearch.SearchTypes.SearchResultItem>().Returns((new[] { searchResultItem }).AsQueryable());
 
-          // act
           Sitecore.Data.Items.Item result = index.CreateSearchContext().GetQueryable<Sitecore.ContentSearch.SearchTypes.SearchResultItem>().Single().GetItem();
 
-          // assert
           result.Paths.FullPath.Should().Be("/sitecore/content/home");
         }
       }
@@ -149,13 +141,10 @@
     [Fact]
     public void HowDoIMockTrackerVisitor()
     {
-      // arrange
       var visitor = Substitute.For<Sitecore.Analytics.Data.DataAccess.Visitor>(Guid.NewGuid());
 
-      // act
       using (new Sitecore.Common.Switcher<Sitecore.Analytics.Data.DataAccess.Visitor>(visitor))
       {
-        // assert
         Sitecore.Analytics.Tracker.Visitor.Should().Be(visitor);
       }
     }